// packages/esql/types.ts
export type ESQLAstCommand =
  | ESQLCommand
  | ESQLAstJoinCommand
  | ESQLAstChangePointCommand
  | ESQLAstRerankCommand
  | ESQLAstCompletionCommand
  | ESQLAstFuseCommand
  | ESQLAstForkCommand
  | ESQLAstUriPartsCommand
  | ESQLAstTsInfoCommand
  | ESQLAstMetricsInfoCommand
  | ESQLAstUserAgentCommand
  | ESQLAstIpLocationCommand
  | ESQLAstRegisteredDomainCommand;

export type ESQLAstAllCommands = ESQLAstCommand | ESQLAstHeaderCommand;

export type ESQLAstNode = ESQLAstCommand | ESQLAstHeaderCommand | ESQLAstExpression | ESQLAstItem;

/**
 * Represents an *expression* in the AST.
 */
export type ESQLAstExpression = ESQLSingleAstItem;

export type ESQLSingleAstItem =
  | ESQLAstQueryExpression
  | ESQLFunction
  | ESQLCommandOption
  | ESQLSource
  | ESQLParens
  | ESQLColumn
  | ESQLDatePeriodLiteral
  | ESQLTimeDurationLiteral
  | ESQLList
  | ESQLLiteral
  | ESQLIdentifier
  | ESQLInlineCast
  | ESQLOrderExpression
  | ESQLUnknownItem
  | ESQLMap
  | ESQLMapEntry
  | PromQLAstQueryExpression;
// ...
export type ESQLAstItem = ESQLSingleAstItem | ESQLAstItem[];
// ...
export interface ESQLLocation {
  min: number;
  max: number;
}

export interface ESQLAstBaseItem<Name = string> {
  name: Name;
  text: string;
  location: ESQLLocation;
  incomplete: boolean;
  formatting?: ESQLAstNodeFormatting;
}

/**
 * Contains optional formatting information used by the pretty printer.
 */
export interface ESQLAstNodeFormatting {
  top?: ESQLAstComment[];
  left?: ESQLAstCommentMultiLine[];
  right?: ESQLAstCommentMultiLine[];
  rightSingleLine?: ESQLAstCommentSingleLine;
  bottom?: ESQLAstComment[];
}

export interface ESQLCommand<Name = string> extends ESQLAstBaseItem<Name> {
  type: 'command';

  /**
   * The subtype of the command. For example, the `JOIN` command can be: (1)
   * LOOKUP JOIN, (2) LEFT JOIN, (3) RIGHT JOIN.
   */
  commandType?: string;

  args: ESQLAstItem[];
}

export interface ESQLAstJoinCommand extends ESQLCommand<'join'> {
  commandType: 'lookup' | 'left' | 'right';
}

export interface ESQLAstChangePointCommand extends ESQLCommand<'change_point'> {
  value: ESQLColumn;
  key?: ESQLColumn;
  target?: {
    type: ESQLColumn;
    pvalue: ESQLColumn;
  };
}

export interface ESQLAstCompletionCommand extends ESQLCommand<'completion'> {
  prompt: ESQLAstExpression;
  inferenceId: ESQLLiteral;
  targetField?: ESQLColumn;
}

export interface ESQLAstFuseCommand extends ESQLCommand<'fuse'> {
  fuseType?: ESQLIdentifier;
}

export interface ESQLAstRerankCommand extends ESQLCommand<'rerank'> {
  query: ESQLLiteral;
  fields: ESQLAstField[];
  targetField?: ESQLColumn;
  inferenceId: ESQLLiteral | undefined;
}

export interface ESQLAstForkCommand extends ESQLCommand<'fork'> {
  args: ESQLForkParens[];
}

export interface ESQLAstMmrCommand extends ESQLCommand<'mmr'> {
  queryVector?: ESQLSingleAstItem;
  diversifyField: ESQLSingleAstItem;
  limit: ESQLSingleAstItem;
  namedParameters?: ESQLSingleAstItem;
}

export interface ESQLAstUriPartsCommand extends ESQLCommand<'uri_parts'> {
  targetField: ESQLColumn;
  expression?: ESQLAstExpression;
}

export interface ESQLAstTsInfoCommand extends ESQLCommand<'ts_info'> {}

export interface ESQLAstMetricsInfoCommand extends ESQLCommand<'metrics_info'> {}

export interface ESQLAstUserAgentCommand extends ESQLCommand<'user_agent'> {
  targetField: ESQLColumn;
  expression?: ESQLAstExpression;
  namedParameters?: ESQLSingleAstItem;
}

export interface ESQLAstIpLocationCommand extends ESQLCommand<'ip_location'> {
  targetField: ESQLColumn;
  expression?: ESQLAstExpression;
  namedParameters?: ESQLSingleAstItem;
}

export interface ESQLAstRegisteredDomainCommand extends ESQLCommand<'registered_domain'> {
  targetField: ESQLColumn;
  expression?: ESQLAstExpression;
}
// ...
export interface ESQLAstHeaderCommand<
  Name extends string = string,
  Arg = ESQLAstExpression,
> extends ESQLAstBaseItem {
  type: 'header-command';

  /** Name of the command */
  name: Name;

  /**
   * Represents the arguments for the command. It has to be a list, because
   * even the SET command was initially designed to accept multiple
   * assignments.
   *
   * Example:
   *
   * ```
   * SET setting1 = "value1", setting2 = "value2"
   * ```
   */
  args: Arg[];
}

export type ESQLAstSetHeaderCommand = ESQLAstHeaderCommand<
  'set',
  ESQLBinaryExpression<BinaryExpressionAssignmentOperator>
>;

export type ESQLIdentifierOrParam = ESQLIdentifier | ESQLParamLiteral;

export interface ESQLCommandOption extends ESQLAstBaseItem {
  type: 'option';
  args: ESQLAstItem[];
}

export interface ESQLAstQueryExpression extends ESQLAstBaseItem<''> {
  type: 'query';
  header?: ESQLAstHeaderCommand[];
  commands: ESQLAstCommand[];
}
// ...
export type FunctionSubtype =
  | 'variadic-call' // fn(a, b, c, ...)
  | 'unary-expression' // -a, +a, NOT a, ...
  | 'postfix-unary-expression' // a IS NULL, a IS NOT NULL, ...
  | 'binary-expression'; // a + b, a - b, a * b, ...

export interface ESQLFunction<
  Subtype extends FunctionSubtype = FunctionSubtype,
  Name extends string = string,
> extends ESQLAstBaseItem<Name> {
  type: 'function';

  /**
   * Default is 'variadic-call'.
   */
  subtype?: Subtype;

  /**
   * A node representing the function or operator being called.
   */
  operator?: ESQLIdentifier | ESQLParamLiteral;

  args: ESQLAstItem[];
}
// ...
export interface ESQLColumn extends ESQLAstBaseItem {
  type: 'column';

  /**
   * Optional qualifier for the column, e.g. index name or alias.
   *
   * @example
   *
   * ```esql
   * [index].[column]
   * [index].[nested.column.part]
   * ```
   *
   * `index` is the qualifier.
   */
  qualifier?: ESQLIdentifier;

  /**
   * A ES|QL column name can be composed of multiple parts,
   * e.g: part1.part2.`part``3️⃣`.?param. Where parts can be quoted, or not
   * quoted, or even be a parameter.
   *
   * The args list contains the parts of the column name.
   */
  args: Array<ESQLIdentifier | ESQLParam>;

  /**
   * An identifier can be composed of multiple parts, e.g: part1.part2.`part``3️⃣`.
   * This property contains the parsed unquoted parts of the identifier.
   * For example: `['part1', 'part2', 'part`3️⃣']`.
   */
  parts: string[];

  /**
   * @deprecated
   *
   * An identifier can be composed of multiple parts, e.g: part1.part2.`part3️⃣`
   *
   * Each part can be quoted or not quoted independently. A single `quoted`
   * property is not enough to represent the identifier. Use `parts` instead.
   */
  quoted: boolean;
}
// ...
export interface ESQLParamLiteral<
  ParamType extends string = string,
  ParamKind extends ESQLParamKinds = ESQLParamKinds,
> extends ESQLAstBaseItem {
  type: 'literal';
  literalType: 'param';
  paramKind: ParamKind;
  paramType: ParamType;
  value: string | number;
}

export type ESQLParamKinds = '?' | '??';
// ...
export type ESQLParam =
  | ESQLUnnamedParamLiteral
  | ESQLNamedParamLiteral
  | ESQLPositionalParamLiteral;

export interface ESQLIdentifier extends ESQLAstBaseItem {
  type: 'identifier';
}